signetapi/requirements.txt
Django==3.2.13
django-cors-headers==3.12.0
django-ratelimit==3.0.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djongo==1.3.6
mysqlclient==2.1.0
Pillow==9.1.1
pillow-avif-plugin==1.3.1
web3
moralis
requests==2.27.1
psycopg2-binary